client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["heyul <heyul.chavez@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
opencv-python = "^4.9.0.80"
numpy = "^1.24.0"
requests = "^2.31.0"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
tzdata = "^2024.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
